requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.0; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
